# Conflicts:
#	Sources/TokamakCore/MountedViews/MountedElement.swift
#	Sources/TokamakCore/Views/Navigation/NavigationView.swift
#	Sources/TokamakCore/Views/Text/TextField.swift
#	Sources/TokamakDOM/Views/Buttons/Button.swift
#	Sources/TokamakDOM/Views/Text/TextField.swift
#	Sources/TokamakDemo/TokamakDemo.swift
#	Sources/TokamakGTK/Modifiers/WidgetModifier.swift

+++ Sources/TokamakCore/Shapes/ModifiedShapes.swift
@@ -1,4 +1,4 @@
-// Copyright 2020 Tokamak contributors
+// Copyright 2020-2021 Tokamak contributors
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -14,6 +14,8 @@
 //
 //  Created by Carson Katri on 06/28/2020.
 //
+
+import Foundation
 
 public struct _StrokedShape<S>: Shape where S: Shape {
   @Environment(\.self) public var environment
@@ -106,7 +108,7 @@ public struct RotatedShape<Content>: Shape where Content: Shape {
   public func path(in rect: CGRect) -> Path {
     shape
       .path(in: rect)
-      .applying(.init(rotationAngle: angle.radians))
+      .applying(.init(rotationAngle: CGFloat(angle.radians)))
   }
 }
